deal with interior mutability
