+++ polars/polars-lazy/polars-pipe/src/executors/sinks/utils.rs
@@ -2,6 +2,7 @@ use polars_core::export::ahash::RandomState;
 use polars_core::prelude::*;
 
 pub(super) fn hash_series(columns: &[Series], buf: &mut Vec<u64>, hb: &RandomState) {
+    debug_assert!(buf.is_empty());
     let mut col_iter = columns.iter();
     let first_key = col_iter.next().unwrap();
     first_key.vec_hash(hb.clone(), buf).unwrap();
